requirements.txt
fastapi
uvicorn
python-multipart
aiogram>=3.0.0
google-generativeai
pydantic
python-dotenv
requests